dev(memory): merge all header as implement and zones handling on boot allocator
