+++ src/test/test.c
@@ -98,9 +98,9 @@ void test_boot_alocator(void)
 	vga_printf("---------------------Testing Boot Allocator---------------------\n");
 	boot_allocator_printer();
 	vga_printf("-----------------------------------------------------------------\n");
-	void *a = boot_alloc(64);
-	void *b = boot_alloc(128);
-	void *c = boot_alloc(256);
+	void *a = boot_alloc(64, HIGHMEM_ZONE);
+	void *b = boot_alloc(128, HIGHMEM_ZONE);
+	void *c = boot_alloc(256, HIGHMEM_ZONE);
 
 	vga_printf("boot_alloc(64) , start = %p | end %p\n", a, a + 64);
 	vga_printf("boot_alloc(128) , start = %p | end %p\n", b, b + 128);
